pyproject.toml
[project]
name = "workbench"
version = "2.1.0"
requires-python = ">=3.10"

[project.optional-dependencies]
test = ["pytest>=8.0", "pyyaml>=6.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
